sentinel_hub/main.py
import os
import datetime
from flask import Flask, render_template, jsonify, request
from google.cloud import firestore

app = Flask(__name__)

# Initialize Firestore (assuming implicit credentials or local emulator)
# db = firestore.Client() 

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/api/status', methods=['GET'])
def get_status():
    # Mock data for now, eventually pull from Firestore
    return jsonify({
        "system_fidelity": 85,
        "phase": "Ignition",
        "active_agents": ["Proposer", "Solver", "Observer"],
        "critical_constraints": {
            "gallium_supply": "Stable",
            "licensing": "Expiring Soon"
        }
    })

@app.route('/api/log', methods=['POST'])
def log_event():
    data = request.json
    # db.collection('logs').add(data)
    print(f"Log received: {data}")
    return jsonify({"status": "logged"}), 201

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=8080, debug=True)
